# shared/api/order/v1/order.openapi.v1.yaml
openapi: 3.1.0
info:
  title: OrderService API
  version: 1.0.0
  description: API for create and manage orders
x-ogen:
  target: ../../../pkg/openapi/order/v1
  package: order_v1
  clean: true

paths:
  /api/v1/orders: # create new order
    post:
      summary: Create a new order with selected details
      operationId: CreateNewOrder
      tags:
      - Order
      requestBody:
        required: true
        content:
          application/json:
            schema:      
              type: object
              required:
                - user_uuid
                - part_uuids
              properties:
                user_uuid:
                  type: string
                  description: user uuid
                  examples: 
                  - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                part_uuids:
                  type: array
                  items:
                    type: string
                  description: uuid of the detail(-s)
                  examples: 
                  - "f47ac10b-58cc-4372-a567-0e02b2c3d479"

      responses:
        '201':
          description: The order created successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - order_uuid
                  - total_price
                properties:
                  order_uuid:
                    type: string
                    description: uuid of the current order
                    examples: 
                      - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  total_price:
                    type: number
                    format: float
                    description: total price of the current order
                    examples: 
                      - 123.45
        '422':
          description: The order can't be created because one of detail is unavalable
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples:
                      - 422
                  message:
                    type: string
                    description: Error description
                    examples: 
                    - "Unprocessable entity error: detail not exists" 
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples: 
                      - 400
                  message:
                    type: string
                    description: Error description
                    examples: 
                    - "Bad Request: Invalid parameter format" 
        '403':
          description: Bad gateway - service is unavalable
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples: 
                      - 403
                  message:
                    type: string
                    description: Error description
                    examples: 
                    - "Bad gateway: service is unavalable" 
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples: 
                      - 500
                  message:
                    type: string
                    description: Error description
                    examples:
                    - "Internal server error occurred" 
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                  message:
                    type: string
                    description: Error description
  /api/v1/orders/{order_uuid}/pay: # pay for order by uuid
    post:
      summary: Current order payment
      operationId: OrderPayment
      tags:
        - PayOrder
      parameters:
      - name: order_uuid
        in: path
        required: true
        description: uuid of the order
        schema:
          type: string
          minLength: 36
          maxLength: 36
          examples: 
          - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_method
              properties:
                payment_method:
                  type: string
                  description: payment method of the order
                  enum: [UNKNOWN, CARD, SBP, TRANSFER]
                  examples: 
                  - "CARD"

      responses:
        '200':
          description: Payment completed successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - transaction_uuid
                properties:
                  transaction_uuid:
                    type: string
                    description: uuid of current transaction
                    examples: 
                    - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples: 
                      - 404
                  message:
                    type: string
                    description: Error description
                    examples: 
                    - "Order is not found" 
        '400':
          description: Bad request - validation error
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples:
                      - 400
                  message:
                    type: string
                    description: Error description
                    examples:
                    - "Bad Request: Invalid parameter format" 
        '403':
          description: Bad gateway - service is unavalable
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples: 
                      - 403
                  message:
                    type: string
                    description: Error description
                    examples: 
                    - "Bad gateway: service is unavalable" 
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples:
                      - 500
                  message:
                    type: string
                    description: Error description
                    examples:
                    - "Internal server error occurred" 
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                  message:
                    type: string
                    description: Error description
  /api/v1/orders/{order_uuid}: # show order by uuid
    get:
      summary: returns order using uuid
      operationId: OrderByUUID
      tags:
        - OrderUUID
      parameters:
      - name: order_uuid
        in: path
        required: true
        description: uuid of the order
        schema:
          type: string
          minLength: 36
          maxLength: 36
          examples: 
            - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        '200':
          description: searching of the order completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  order_uuid:
                    type: string
                    description: order uuid
                    examples:
                      - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  user_uuid:
                    type: string
                    description: user uuid
                    examples:
                      - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  part_uuids:
                    type: array
                    items:
                      type: string
                    description: uuid of the detail(-s)
                    examples:
                      - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  total_price:
                    type: number
                    format: double
                    description: total price of the order
                    examples:
                      - 123.45
                  transaction_uuid:
                    type: string
                    description: transaction uuid
                    examples:
                      - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                  payment_method:
                    type: string
                    description: payment method of the order
                    enum: [UNKNOWN, CARD, SBP, TRANSFER]
                    examples:
                      - "CARD"
                  status:
                    type: string
                    description: order payment status
                    enum: [PANDING_PAYMENT, PAID, CANCELLED]
        '404':
          description: order is not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples:
                      - 404
                  message:
                    type: string
                    description: Error description
                    examples:
                      - "Order is not found" 
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                  message:
                    type: string
                    description: Error description
  /api/v1/orders/{order_uuid}/cancel: # cancel order by uuid (before payment)
    post:
      summary: Cancel order by uuid
      operationId: OrderCancel
      tags:
      - OrderCancel
      parameters:
      - name: order_uuid
        in: path
        required: true
        description: uuid of the order
        schema:
          type: string
          minLength: 36
          maxLength: 36
          examples:
          - "f47ac10b-58cc-4372-a567-0e02b2c3d479"
      responses:
        '204':
          description: order canceled successfully
        '404':
          description: order is not found
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples: 
                      - 404
                  message:
                    type: string
                    description: Error description
                    examples:
                    - "Order is not found"
        '409':
          description: the order already is paid and can't be cancelled
          content:
            application/json:
              schema:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                    examples:
                      - 404
                  message:
                    type: string
                    description: Error description
                    examples:
                    - "Order is paid, can't be cancelled"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    description: HTTP-code error
                  message:
                    type: string
                    description: Error description